Move the crate-name-fixing script into Python
I wanted the utiltity functions of this repository to be written in Rust,
but the fact is that having a program that fixes crate name/path errors
be dependent on the crate name/path system having no errors because it
exists in the same cargo workspace is a losing battle. It will be much
more streamlined if the gen-workspace script completely bypasses cargo
by running in python.